评论功能初步整合

+++ src/main/resources/db/migration/V3__Create_comment_table.sql
@@ -3,7 +3,7 @@ create table comment
     id BIGINT auto_increment primary key ,
     pid BIGINT null comment '父类评论id',
     type tinyint null comment '父类类型',
-    commentator int null comment '评论人id',
+    commentator BIGINT null comment '评论人id',
     like_count BIGINT default 0 null,
     gmt_modified BIGINT null,
     gmt_create BIGINT null,
